requirements.txt
graphviz==0.15
seaborn==0.11.2
matplotlib==3.5.1
streamlit==1.5.1
pandas==1.4.3
pm4py==2.2.27
missingno==0.4.2
plotly==4.14.1
pydotplus==2.0.2
numpy==1.23.1
pyvis==0.2.1
Pillow==9.2.0
